Cleanup: move bmqsys_mocktime -> bmqtst_mocktime
