+++ src/groups/bmq/bmqimp/package/bmqimp.dep
@@ -6,6 +6,5 @@ bmqp
 bmqpi
 bmqscm
 bmqst
-bmqsys
 bmqt
 bmqu
