+++ src/groups/bmq/bmqio/package/bmqio.dep
@@ -1,7 +1,6 @@
 bmqex
 bmqscm
 bmqst
-bmqsys
 bmqt
 bmqu
 bmqvt
