+++ src/groups/bmq/bmqsys/doc/bmqsys.txt
@@ -9,20 +9,16 @@ components to be reused through various applications.
 
 /Hierarchical Synopsis
 /---------------------
-The 'bmqsys' package currently has 3 components having 2 levels of physical
+The 'bmqsys' package currently has 2 components having 2 levels of physical
 dependency.  The list below shows the hierarchal ordering of the components.
 ..
-  2. bmqsys_mocktime
-     bmqsys_statmonitorsnapshotrecorder
+  2. bmqsys_statmonitorsnapshotrecorder
 
   1. bmqsys_statmonitor
 ..
 
 /Component Synopsis
 /------------------
-: 'bmqsys_mocktime:
-:      Provide a mock of the time accessors usable with 'bmqsys::Time'.
-:
 : 'bmqsys_statmonitor':
 :      Provide a mechanism to monitor cpu and memory stats.
 :
