+++ src/groups/bmq/bmqsys/package/bmqsys.mem
@@ -1,3 +1,2 @@
-bmqsys_mocktime
 bmqsys_statmonitor
 bmqsys_statmonitorsnapshotrecorder
