+++ src/groups/bmq/bmqtst/doc/bmqtst.txt
@@ -9,10 +9,11 @@ writing test drivers.
 
 /Hierarchical Synopsis
 /---------------------
- The bmqtst' package currently has 3 components having 1 level of physical
+ The bmqtst' package currently has 4 components having 1 level of physical
  dependency.  The list below shows the hierarchical ordering of the components.
 ..
   1. bmqtst_blobtestutil
+     bmqtst_mocktime
      bmqtst_scopedlogobserver
      bmqtst_testhelper
 ..
@@ -21,6 +22,8 @@ writing test drivers.
 /------------------
 : 'bmqtst_blobtestutil':
 :      Provide blob utilities for use in test drivers.
+: 'bmqtst_mocktime':
+:      Provide a mock of the time accessors usable with 'bmqu::Time'.
 : 'bmqtst_scopedlogobserver':
 :      Provide a scoped tester implementation of the log observer protocol.
 : 'bmqtst_testhelper':
